+++ src/main/java/com/wootecam/jspcafe/listener/DependencyListener.java
@@ -3,9 +3,9 @@ package com.wootecam.jspcafe.listener;
 import com.wootecam.jspcafe.repository.UserRepository;
 import com.wootecam.jspcafe.service.UserService;
 import com.wootecam.jspcafe.servlet.HomeServlet;
-import com.wootecam.jspcafe.servlet.SignupFormServlet;
-import com.wootecam.jspcafe.servlet.UserProfileServlet;
-import com.wootecam.jspcafe.servlet.UserServlet;
+import com.wootecam.jspcafe.servlet.user.SignupFormServlet;
+import com.wootecam.jspcafe.servlet.user.UserProfileServlet;
+import com.wootecam.jspcafe.servlet.user.UserServlet;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletContextEvent;
 import jakarta.servlet.ServletContextListener;
